Update suspended cleaning command (#10195)
closes https://github.com/twentyhq/core-team-issues/issues/382

+++ packages/twenty-server/src/engine/core-modules/workspace/services/workspace.service.ts
@@ -14,6 +14,7 @@ import { Repository } from 'typeorm';
 import { BillingEntitlementKey } from 'src/engine/core-modules/billing/enums/billing-entitlement-key.enum';
 import { BillingSubscriptionService } from 'src/engine/core-modules/billing/services/billing-subscription.service';
 import { BillingService } from 'src/engine/core-modules/billing/services/billing.service';
+import { CustomDomainService } from 'src/engine/core-modules/domain-manager/services/custom-domain.service';
 import { DomainManagerService } from 'src/engine/core-modules/domain-manager/services/domain-manager.service';
 import { EnvironmentService } from 'src/engine/core-modules/environment/environment.service';
 import { ExceptionHandlerService } from 'src/engine/core-modules/exception-handler/exception-handler.service';
@@ -35,9 +36,9 @@ import {
   PermissionsExceptionMessage,
 } from 'src/engine/metadata-modules/permissions/permissions.exception';
 import { PermissionsService } from 'src/engine/metadata-modules/permissions/permissions.service';
+import { WorkspaceCacheStorageService } from 'src/engine/workspace-cache-storage/workspace-cache-storage.service';
 import { WorkspaceManagerService } from 'src/engine/workspace-manager/workspace-manager.service';
 import { DEFAULT_FEATURE_FLAGS } from 'src/engine/workspace-manager/workspace-sync-metadata/constants/default-feature-flags';
-import { CustomDomainService } from 'src/engine/core-modules/domain-manager/services/custom-domain.service';
 
 @Injectable()
 // eslint-disable-next-line @nx/workspace-inject-workspace-repository
@@ -61,6 +62,7 @@ export class WorkspaceService extends TypeOrmQueryService<Workspace> {
     private readonly exceptionHandlerService: ExceptionHandlerService,
     private readonly permissionsService: PermissionsService,
     private readonly customDomainService: CustomDomainService,
+    private readonly workspaceCacheStorageService: WorkspaceCacheStorageService,
   ) {
     super(workspaceRepository);
   }
@@ -259,43 +261,71 @@ export class WorkspaceService extends TypeOrmQueryService<Workspace> {
     });
   }
 
-  async softDeleteWorkspace(id: string) {
-    const workspace = await this.workspaceRepository.findOneBy({ id });
-
-    assert(workspace, 'Workspace not found');
-
-    await this.userWorkspaceRepository.delete({ workspaceId: id });
+  async deleteMetadataSchemaCacheAndUserWorkspace(workspace: Workspace) {
+    await this.userWorkspaceRepository.delete({ workspaceId: workspace.id });
 
     if (this.billingService.isBillingEnabled()) {
       await this.billingSubscriptionService.deleteSubscriptions(workspace.id);
     }
 
-    await this.workspaceManagerService.delete(id);
+    await this.workspaceManagerService.delete(workspace.id);
 
     return workspace;
   }
 
-  async deleteWorkspace(id: string) {
-    const userWorkspaces = await this.userWorkspaceRepository.findBy({
-      workspaceId: id,
+  async deleteWorkspace(id: string, softDelete = false) {
+    const workspace = await this.workspaceRepository.findOne({
+      where: { id },
+      withDeleted: true,
     });
 
-    const workspace = await this.softDeleteWorkspace(id);
+    assert(workspace, 'Workspace not found');
+
+    const userWorkspaces = await this.userWorkspaceRepository.find({
+      where: {
+        workspaceId: id,
+      },
+      withDeleted: true,
+    });
 
     for (const userWorkspace of userWorkspaces) {
-      await this.handleRemoveWorkspaceMember(id, userWorkspace.userId);
+      await this.handleRemoveWorkspaceMember(
+        id,
+        userWorkspace.userId,
+        softDelete,
+      );
     }
 
-    await this.workspaceRepository.delete(id);
+    await this.workspaceCacheStorageService.flush(
+      workspace.id,
+      workspace.metadataVersion,
+    );
 
-    return workspace;
+    if (softDelete) {
+      return await this.workspaceRepository.softDelete({ id });
+    }
+
+    await this.deleteMetadataSchemaCacheAndUserWorkspace(workspace);
+
+    return await this.workspaceRepository.delete(id);
   }
 
-  async handleRemoveWorkspaceMember(workspaceId: string, userId: string) {
-    await this.userWorkspaceRepository.delete({
-      userId,
-      workspaceId,
-    });
+  async handleRemoveWorkspaceMember(
+    workspaceId: string,
+    userId: string,
+    softDelete = false,
+  ) {
+    if (softDelete) {
+      await this.userWorkspaceRepository.softDelete({
+        userId,
+        workspaceId,
+      });
+    } else {
+      await this.userWorkspaceRepository.delete({
+        userId,
+        workspaceId,
+      });
+    }
 
     const userWorkspaces = await this.userWorkspaceRepository.find({
       where: {
